added service tests

+++ internal/service/swift_service.go
@@ -3,13 +3,15 @@ package service
 import (
 	"SWIFT_task/internal/model"
 	"SWIFT_task/internal/repository"
+	"fmt"
+	"strconv"
 )
 
 type BranchService struct {
-	branchRepo *repository.BranchRepository
+	branchRepo repository.BranchRepository
 }
 
-func NewBranchService(branchRepo *repository.BranchRepository) *BranchService {
+func NewBranchService(branchRepo repository.BranchRepository) *BranchService {
 	return &BranchService{branchRepo: branchRepo}
 }
 
@@ -19,7 +21,7 @@ func (s *BranchService) GetBranchDetails(swiftCode string) (*model.BranchDto, er
 		return nil, err
 	}
 	if branch == nil {
-		return nil, nil
+		return nil, NewBranchNotExistsError("Branch with swiftCode " + swiftCode + " not exists.")
 	}
 
 	branchDto := branch.ToBranchDto()
@@ -42,6 +44,10 @@ func (s *BranchService) GetBranchDetails(swiftCode string) (*model.BranchDto, er
 }
 
 func (s *BranchService) GetBranchesByISO2code(countryISO2code string) (*model.BranchesForCountryDto, error) {
+	if len(countryISO2code) != 2 {
+		return nil, NewValidationError("Len of countryISO2 code should be 2. Got len: " + strconv.Itoa(len(countryISO2code)))
+	}
+
 	branches, err := s.branchRepo.GetBranchesByISO2code(countryISO2code)
 	if err != nil {
 		return nil, err
@@ -63,14 +69,30 @@ func (s *BranchService) GetBranchesByISO2code(countryISO2code string) (*model.Br
 	}, err
 }
 
-func (s *BranchService) AddSwiftCode(branch *model.Branch) error {
-	return s.branchRepo.CreateBranch(branch)
-}
+func (s *BranchService) AddSwiftCode(branch model.Branch) error {
+	if len(branch.CountryISO2) != 2 {
+		return NewValidationError("Len of countryISO2 code should be 2. Got len: " + strconv.Itoa(len(branch.CountryISO2)))
+	}
 
-func (s *BranchService) RemoveBranchBySwiftCode(swiftCode string) error {
-	_, err := s.branchRepo.GetBranchBySwiftCode(swiftCode)
+	existingBranch, err := s.branchRepo.GetBranchBySwiftCode(branch.SwiftCode)
 	if err != nil {
 		return err
 	}
+	if existingBranch != nil {
+		return NewBranchExistsError("Branch with swiftCode " + branch.SwiftCode + " already exists.")
+	}
+	fmt.Println("Adding swift code: " + branch.SwiftCode)
+	return s.branchRepo.CreateBranch(&branch)
+}
+
+func (s *BranchService) RemoveBranchBySwiftCode(swiftCode string) error {
+	branch, err := s.branchRepo.GetBranchBySwiftCode(swiftCode)
+	if err != nil {
+		return err
+	}
+	if branch == nil {
+		return NewBranchNotExistsError("Branch with swiftCode " + swiftCode + " not exists.")
+	}
+
 	return s.branchRepo.RemoveBranchBySwiftCode(swiftCode)
 }
